refactor tests (some broken - WIP)
