+++ tests/k8s_cluster_test.py
@@ -31,7 +31,7 @@ from hpecp.k8s_cluster import (
 )
 
 from .base_test import BaseTestCase, MockResponse, get_client
-from .k8s_cluster_mock_api import mockApiSetup
+from .k8s_cluster_mock_api_responses import mockApiSetup
 
 # setup the mock data
 mockApiSetup()
