+++ tests/k8s_worker_test.py
@@ -13,136 +13,10 @@ from .base_test import (
     get_client,
     session_mock_response,
 )
+from .k8s_worker_mock_api_responses import mockApiSetup
 
-BaseTestCase.registerHttpGetHandler(
-    url="https://127.0.0.1:8080/api/v2/worker/k8shost",
-    response=MockResponse(
-        json_data={
-            "_embedded": {
-                "k8shosts": [
-                    {
-                        "status": "unlicensed",
-                        "propinfo": {
-                            "bds_storage_apollo": "false",
-                            "bds_network_publicinterface": "ens5",
-                        },
-                        "approved_worker_pubkey": [],
-                        "tags": [],
-                        "hostname": (
-                            "ip-10-1-0-238.eu-west-2.compute.internal"
-                        ),
-                        "ipaddr": "10.1.0.238",
-                        "setup_log": (
-                            "/var/log/bluedata/install/"
-                            "k8shost_setup_10.1.0.238-"
-                            "2020-4-26-18-41-16"
-                        ),
-                        "_links": {
-                            "self": {"href": "/api/v2/worker/k8shost/4"}
-                        },
-                        "sysinfo": {
-                            "network": [],
-                            "keys": {
-                                "reported_worker_public_key": (
-                                    "ssh-rsa ...== server\n"
-                                )
-                            },
-                            "storage": [],
-                            "swap": {"swap_total": 0},
-                            "memory": {"mem_total": 65842503680},
-                            "gpu": {"gpu_count": 0},
-                            "cpu": {
-                                "cpu_logical_cores": 16,
-                                "cpu_count": 8,
-                                "cpu_physical_cores": 8,
-                                "cpu_sockets": 1,
-                            },
-                            "mountpoint": [],
-                        },
-                    },
-                    {
-                        "status": "bundle",
-                        "approved_worker_pubkey": [],
-                        "tags": [],
-                        "hostname": "",
-                        "ipaddr": "10.1.0.186",
-                        "setup_log": (
-                            "/var/log/bluedata/install/"
-                            "k8shost_setup_10.1.0.186-"
-                            "2020-4-26-18-49-10"
-                        ),
-                        "_links": {
-                            "self": {"href": "/api/v2/worker/k8shost/5"}
-                        },
-                    },
-                ]
-            }
-        },
-        status_code=200,
-        headers={},
-    ),
-)
-
-BaseTestCase.registerHttpGetHandler(
-    url="https://127.0.0.1:8080/api/v2/worker/k8shost/5",
-    response=MockResponse(
-        json_data={
-            "status": "bundle",
-            "approved_worker_pubkey": [],
-            "tags": [],
-            "hostname": "",
-            "ipaddr": "10.1.0.186",
-            "setup_log": (
-                "/var/log/bluedata/install/"
-                "k8shost_setup_10.1.0.186-"
-                "2020-4-26-18-49-10"
-            ),
-            "_links": {"self": {"href": "/api/v2/worker/k8shost/5"}},
-        },
-        status_code=200,
-        headers={},
-    ),
-)
-
-BaseTestCase.registerHttpGetHandler(
-    url="https://127.0.0.1:8080/api/v2/worker/k8shost/5?setup_log=true",
-    response=MockResponse(
-        json_data={
-            "status": "bundle",
-            "approved_worker_pubkey": [],
-            "tags": [],
-            "hostname": "",
-            "ipaddr": "10.1.0.186",
-            "setup_log": (
-                "/var/log/bluedata/install/"
-                "k8shost_setup_10.1.0.186-"
-                "2020-4-26-18-49-10"
-            ),
-            "_links": {"self": {"href": "/api/v2/worker/k8shost/5"}},
-        },
-        status_code=200,
-        headers={},
-    ),
-)
-
-BaseTestCase.registerHttpGetHandler(
-    url="https://127.0.0.1:8080/api/v2/worker/k8shost/8",
-    response=MockResponse(
-        json_data={}, status_code=404, raise_for_status_flag=True, headers={},
-    ),
-)
-
-BaseTestCase.registerHttpPostHandler(
-    url="https://127.0.0.1:8080/api/v2/worker/k8shost/5",
-    response=MockResponse(json_data={}, status_code=204, headers={}),
-)
-
-BaseTestCase.registerHttpPostHandler(
-    url="https://127.0.0.1:8080/api/v2/worker/k8shost/",
-    response=MockResponse(
-        json_data={}, status_code=201, headers={"location": "/new/cluster/id"},
-    ),
-)
+# setup the mock data
+mockApiSetup()
 
 
 class TestWorkers(BaseTestCase):
